% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateHouseholdTable.R
\name{add_dilution_factor}
\alias{add_dilution_factor}
\title{Add dilution factor}
\usage{
add_dilution_factor(df, dilution_factor)
}
\arguments{
\item{df}{The data frame the dilution factor should be added on.
Must contain the column "Sample description 1".}

\item{dilution_factor}{A named vector that matches sample description (name)
with the used dilution factor (value).}
}
\value{
Data frame updated with dilution factor
}
\description{
Add the dilution factor to the table in define_groups. Gives warning if input objects do not match
}
