% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractInfos.R
\name{compute_groupwise_mean}
\alias{compute_groupwise_mean}
\title{Compute group-wise mean}
\usage{
compute_groupwise_mean(df, group_by, new_col, old_col)
}
\arguments{
\item{df}{Data frame with experiment information.}

\item{group_by}{Column(s) by which the data should be grouped.}

\item{new_col}{The name of the new column.}

\item{old_col}{The name of the column the mean should be computed from.}
}
\value{
df updated by new_col.
}
\description{
Compute the mean of a specified group of Wells for a certain column and add
as new column
}
