% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateHouseholdTable.R
\name{compute_shearing_factor}
\alias{compute_shearing_factor}
\title{Create RPP30 table}
\usage{
compute_shearing_factor(df, mean_copies_factor, mean_cells_per_reac_factor)
}
\arguments{
\item{df}{data frame containing information on wells with RPP30 and RPP30Shear}

\item{mean_copies_factor}{Number to multiply Mean concentration RPP30 + Shear with to compute mean copies/well}

\item{mean_cells_per_reac_factor}{Factor to multiply Mean copies/cell with to obtain Mean cells per reaction (named list)}
}
\value{
Updated data frame
}
\description{
Create the RPP30 table to compute Shearing factor and mean RPP30(Shear) concentration.
}
