% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractInfos.R
\name{compute_target_means}
\alias{compute_target_means}
\title{Compute target mean}
\usage{
compute_target_means(tab)
}
\arguments{
\item{tab}{Output from merge_tables}
}
\value{
Updated table
}
\description{
This function computes the number of targets (genes) detected per million cells
and the mean thereof within the wells of one group
}
