% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractInfos.R
\name{compute_total_HIV_envPsi}
\alias{compute_total_HIV_envPsi}
\title{Compute total HIV content}
\usage{
compute_total_HIV_envPsi(df)
}
\arguments{
\item{df}{The data frame for whose column the total HIV content shall be computed}
}
\value{
Data frame df updated with new column.
}
\description{
Estimate the total HIV content by Env Psi content.
}
