% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractInfos.R
\name{create_confusion_matrix}
\alias{create_confusion_matrix}
\title{Create a confusion matrix}
\usage{
create_confusion_matrix(df, data_table, ch_dye, target_channel)
}
\arguments{
\item{df}{in_csv subsetted for wells containing targets}

\item{data_table}{dtQC subsetted for wells containing target}

\item{ch_dye}{named list with used channels as names and respective dyes as values}

\item{target_channel}{named list with Target (in_cs columns) as names and channels as values}
}
\value{
Confusion matrix
}
\description{
Create a confusion matrix showing which targets and multiplets (combination of
targets) are found in which wells
}
