% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Readers.R
\name{create_dtQC}
\alias{create_dtQC}
\title{Create QC table}
\usage{
create_dtQC(xlsx_df, cols_of_int = NULL, acc_drop_factor = 3)
}
\arguments{
\item{xlsx_df}{The data frame created when reading the xlsx file.}

\item{cols_of_int}{Columns that should be contained in dtQC. If not specified,
a standard set of columns is kept. (optional)}

\item{acc_drop_factor}{Numeric factor by which the threshold is computed from
the accepted droplets. (default=3)}
}
\value{
The dtQC dataframe with the relevant columns.
}
\description{
Create the dtQC table, which contains the relevant content from the xlsx file.
}
\keyword{internal}
