% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateHouseholdTable.R
\name{create_household_table}
\alias{create_household_table}
\title{Create RPP30 table}
\usage{
create_household_table(
  dtQC,
  grouped_data,
  thresh,
  mean_copies_factor,
  mean_cells_per_reac_factor
)
}
\arguments{
\item{dtQC}{The data frame created in the Readers/read_files function.}

\item{grouped_data}{Data frame created in define_groups function.}

\item{thresh}{Minimum number of accepted droplets to include the well in the calculations.
Other wise it will be removed during the process.}

\item{mean_copies_factor}{Number to multiply Mean concentration RPP30 + Shear with to compute mean copies/well # I guess should be named volume}

\item{mean_cells_per_reac_factor}{Factor to multiply Mean copies/cell with to obtain Mean cells per reaction}
}
\value{
List with two elements: data frame with information about RPP30(Shear) and an updated grouped_data.
}
\description{
Create the RPP30 table to compute Shearing factor and mean RPP30(Shear) concentration.
}
