% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractInfos.R
\name{create_table}
\alias{create_table}
\title{Create analysis table for batch}
\usage{
create_table(
  batch,
  conf_mat,
  num_target,
  ch_dye,
  multi_pos,
  thresh,
  tar_mio_factor,
  tab1
)
}
\arguments{
\item{batch}{The wells which are analysed together.}

\item{conf_mat}{Confusion matrix (well vs. positives) for the batch.}

\item{num_target}{The total number of targets.}

\item{ch_dye}{Named list to match channel (names) with dyes (values).}

\item{multi_pos}{The possible multiple positives.}

\item{thresh}{The minimum number of accepted droplets in well to be included
in the analysis.}

\item{tar_mio_factor}{Factor to multiply Concentration with to obtain
Target/Mio cells}

\item{tab1}{Data frame created in create_household_table().}
}
\value{
List with two elements: a data frame and a sting indicating the nature
of the data sample (empty, H2O control, data)
}
\description{
Combine input data and compute analysis parameters for a group of wells (batch).
}
