% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractInfos.R
\name{create_tables}
\alias{create_tables}
\title{Create analysis tables}
\usage{
create_tables(
  grouped_data,
  in_csv,
  ch_dye,
  multi_pos,
  thresh,
  tar_mio_factor,
  tab1
)
}
\arguments{
\item{grouped_data}{Data frame that gives information per group of wells.
Computed in define_groups()}

\item{in_csv}{Data frame with information from csv input file. Output from 
read_csv().}

\item{ch_dye}{Named list to match channel (names) with dyes (values).}

\item{multi_pos}{The possible multiple positives.}

\item{thresh}{The minimum number of accepted droplets in well to be included
in the analysis.}

\item{tar_mio_factor}{Factor to multiply Concentration with to obtain
Target/Mio cells}

\item{tab1}{Data frame created in create_household_table().}
}
\value{
3 Lists containing the (1) data analysis, (2) confusion matrices, (3)
the H2O samples.
}
\description{
Combine input data and do analysis.
}
