% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateHouseholdTable.R
\name{define_groups}
\alias{define_groups}
\title{Define groups}
\usage{
define_groups(dtQC, dilution_factor)
}
\arguments{
\item{dtQC}{The data frame created in the Readers/read_files function.}

\item{dilution_factor}{The dilution factor will be applied to the concentration
to correct for dilution during the experiment.}
}
\value{
grouped_data data frame, with basic information to each found group.
}
\description{
Define the groups of Well that are analysed together. This is based upon the
columns "Sample description 1" and "Target". If they are the same for two
wells, they will be analysed together.
}
