% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractInfos.R
\name{extract_target_positive_colnames}
\alias{extract_target_positive_colnames}
\title{Extract target positive column names}
\usage{
extract_target_positive_colnames(target, colnames)
}
\arguments{
\item{target}{The target the column names should be positive for.}

\item{colnames}{The column names of a data frame.}
}
\value{
Vector containing the gene positive column names.
}
\description{
Extract those column names that are positive for a specified target (gene).
}
