% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateHouseholdTable.R
\name{get_group_id}
\alias{get_group_id}
\title{Define groups of well to analyse together}
\usage{
get_group_id(dtQC)
}
\arguments{
\item{dtQC}{output from read_files}
}
\value{
Named list mapping well to group
}
\description{
Define groups based on the same sample description and targets found in this well
to analyse the group together.
}
