% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TotalHIVDNA.R
\name{get_intact_multiplet_conc}
\alias{get_intact_multiplet_conc}
\title{Get multiplet column}
\usage{
get_intact_multiplet_conc(df, multiplet, substrings)
}
\arguments{
\item{df}{data table}

\item{multiplet}{number of targets to look for (2 for doublets, 3 for multiplets)}

\item{substrings}{list of targets available}
}
\value{
sum of multiplets
}
\description{
get columns to use for multiplet (triplet or doublet) computation
}
