% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultipPositives.R
\name{get_multi_pos}
\alias{get_multi_pos}
\title{Get multiple positives concentrations}
\usage{
get_multi_pos(df, genes, tar_mio_factor)
}
\arguments{
\item{df}{Data frame with the information on the measurement.}

\item{genes}{The genes for which the sample should be positive
(e.g., c("Env", "Psi") for Env+Psi+}

\item{tar_mio_factor}{Factor to multiply Concentration with to obtain
Target/Mio cells}
}
\value{
Updated data frame (df)
}
\description{
Compute the (intact) concentration and its mean across grouped Well(s) for
multiple positives (e.g., Env+Psi+, Env+Gag+Pol+Psi+).
}
\keyword{internal}
