% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultipPositives.R
\name{get_multiplet_count}
\alias{get_multiplet_count}
\title{Get count for a multiplet}
\usage{
get_multiplet_count(df, genes)
}
\arguments{
\item{df}{data_table}

\item{genes}{list for which multiplets are computed for}
}
\value{
(Sum of) multiplet count
}
\description{
This function identifies the column name(s) of a multiplet based on a target 
list and returns the column values or the sum thereof
}
