% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractInfos.R
\name{get_multipos}
\alias{get_multipos}
\title{Get multiple positive combinations}
\usage{
get_multipos(targets)
}
\arguments{
\item{targets}{List of targets used in the experiment.}
}
\value{
List of possible combinations of input genes
}
\description{
Compute the possible combinations of targets (multiplets) based on the targets
in the experiment.
}
