% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TotalHIVDNA.R
\name{get_quad_pos}
\alias{get_quad_pos}
\title{Get Quad-pos. column}
\usage{
get_quad_pos(df)
}
\arguments{
\item{df}{data table}
}
\value{
quad positive value
}
\description{
get column to use as shear-corrected quadrupel positives with the order 
(quadrupel > EnvGagPsi > EnvPsiPol > EnvPsi > EnvGagPol > EnvGag)
}
