% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractInfos.R
\name{get_subset}
\alias{get_subset}
\title{Get relevant subsets}
\usage{
get_subset(multi_pos)
}
\arguments{
\item{multi_pos}{The multiple positive target.}
}
\value{
List of the possible subsets.
}
\description{
Compute the possible subsets of a multiple positive target.
}
