% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Checks.R
\name{is_dtQC}
\alias{is_dtQC}
\title{Check if df is dtQC}
\usage{
is_dtQC(dtQC)
}
\arguments{
\item{dtQC}{The data frame to be tested.}
}
\value{
Boolean
}
\description{
Check if data frame object fulfills requirements for dtQC.
}
