% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Checks.R
\name{is_grouped_data}
\alias{is_grouped_data}
\title{Check if is grouped data}
\usage{
is_grouped_data(grouped_data)
}
\arguments{
\item{grouped_data}{The object to check}
}
\value{
Boolean
}
\description{
Check if an object is a grouped data df
}
