% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Checks.R
\name{is_named_vector}
\alias{is_named_vector}
\title{several functions that check the correctness of data frames, vectors, input,
and output format etc.
Check if is named vector}
\usage{
is_named_vector(obj)
}
\arguments{
\item{obj}{The object to check}
}
\value{
Boolean
}
\description{
Check if an object is a named vector
}
