% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractInfos.R
\name{match_channel_gene}
\alias{match_channel_gene}
\title{Match measurement channel to gene}
\usage{
match_channel_gene(df, ch_dye, num_ch)
}
\arguments{
\item{df}{Data frame with columns "DyeName(s)" and "Target".}

\item{ch_dye}{Named vector matching channels (names) with dyes (values).}

\item{num_ch}{The number of channels used in total (must be 4 or 5).}
}
\value{
Vector of genes.
}
\description{
Match the Channels that were measured with the genes using the dye information.
}
