% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractInfos.R
\name{merge_tables}
\alias{merge_tables}
\title{Merging information of both input files}
\usage{
merge_tables(data_table, conf_mat, shear_table)
}
\arguments{
\item{data_table}{Information coming from in_xlsx file, wells with targets only}

\item{conf_mat}{Output from create_confusion_matrix}

\item{shear_table}{Output from compute_shearing_factor}
}
\value{
Combined table, conserving all information
}
\description{
This function creates one table based on the two input files: in_xlsx (as dtQC
or data_table in tutorial) and in_csv (as confusion matrix in tutorial).
Merging is based on wells, information from shearing table will be added.
}
