% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Readers.R
\name{read_csv}
\alias{read_csv}
\title{Read csv file}
\usage{
read_csv(filename, csv_skip)
}
\arguments{
\item{filename}{Path to an existing csv file.}

\item{csv_skip}{Number of lines in csv file that should not be read in, since
they are before the actual table. Otherwise reading the file will return an
error due to column number mismatch in different rows.}

\item{csv_end}{Last row number in csv file with relevant data. Otherwise the
following table (in the expected input) will cause an error due to column
number mismatch in different rows.}
}
\value{
Data frame with the relevant data.
}
\description{
Read the csv file of the input files within the specified rows.
}
