% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Readers.R
\name{read_files}
\alias{read_files}
\title{Read files}
\usage{
read_files(
  xlsx_file,
  csv_file,
  csv_skip,
  remove_channel = NULL,
  rm_zero_channels = FALSE
)
}
\arguments{
\item{xlsx_file}{Path to an existing xlsx_file.}

\item{csv_file}{Path to an existing csv_file.}

\item{csv_skip}{Number of lines in csv file that should not be read in, since
they are before the actual table. Otherwise reading the file will return an
error due to column number mismatch in different rows.}

\item{remove_channel}{List of channels that are to be removed. (optional)}

\item{rm_zero_channels}{Boolean value to indicate whether defect channels, 
i.e., channels with zero concentration are to be removed. (optional)}
}
\description{
Read the input xlsx and csv file to create the initial tables for the analysis.
}
