% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Readers.R
\name{read_multiple_files}
\alias{read_multiple_files}
\title{Read multiple files}
\usage{
read_multiple_files(
  xlsx_files,
  csv_files,
  csv_skip = 4,
  remove_channel = NULL,
  rm_zero_channels = FALSE
)
}
\arguments{
\item{xlsx_files}{Vector with paths to existing xlsx_files.}

\item{csv_files}{Vector with paths to existing csv_files.}

\item{csv_skip}{Number of lines in csv file that should not be read in, since
they are before the actual table. Otherwise reading the file will return an
error due to column number mismatch in different rows. Same in every file.}

\item{remove_channel}{List of lists of channels that are to be removed. Will
be processed in the same order as the csv and xlsx files. (optional)}

\item{rm_zero_channels}{Boolean value to indicate whether defect channels, 
i.e., channels with zero concentration are to be removed. (optional)}
}
\description{
Read a list of input xlsx and csv files to create the initial tables for the analysis.
}
