% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Readers.R
\name{read_xlsx}
\alias{read_xlsx}
\title{Read xlsx file}
\usage{
read_xlsx(filename)
}
\arguments{
\item{filename}{Path to an existing xlsx file.}
}
\value{
Data frame with the contents of the xlsx file. NA values in
concentration column are filled in with zeros.
}
\description{
Read the excel file (first sheet) of the input files
}
