% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Readers.R
\name{rm_channels}
\alias{rm_channels}
\title{Remove channels manually (optional)}
\usage{
rm_channels(df, channels)
}
\arguments{
\item{df}{Data frame from which the channels are to be removed. Usually the
output of reading the csv and the xlsx file each.}

\item{channels}{Channels to remove (Well ID)}
}
\value{
Data frame without the specified channels
}
\description{
Remove channels if they are not clean or irrelevant.
}
\keyword{internal}
