% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Readers.R
\name{rm_zero_channel}
\alias{rm_zero_channel}
\title{Remove zero channel}
\usage{
rm_zero_channel(dtQC, in_csv)
}
\arguments{
\item{dtQC}{Data frame created with create_dtQC.}

\item{in_csv}{Data frame created with read_csv.}
}
\value{
List of data frames dtQC and in_csv without the defect channels.
}
\description{
Remove wells for which at least one channel is defect, i.e., has
concentration 0. Wells with H2O probes are not removed.
}
\keyword{internal}
