% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateHouseholdTable.R
\name{sufficient_droplets}
\alias{sufficient_droplets}
\title{Check for sufficient droplets}
\usage{
sufficient_droplets(df, thresh, droplet_col = "Accepted Droplets")
}
\arguments{
\item{df}{The data frame that should be checked. Must contain the columns
"Well" and the one specified in droplet_col.}

\item{thresh}{The threshold (minimum of accepted droplets).}

\item{droplet_col}{The column containing the information about the accepted
droplets (default: "Accepted Droplets")}
}
\value{
Data frame with Wells above the given threshold only.
}
\description{
Check if the number of accepted droplets is above the minimum threshold.
}
