% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractInfos.R
\name{sum_target_positive_values}
\alias{sum_target_positive_values}
\title{Compute total target positive values}
\usage{
sum_target_positive_values(df)
}
\arguments{
\item{df}{Data frame with experiment information.}
}
\value{
Numeric: sum of all target positive values.
}
\description{
Compute sum of all target positive values (including doublets, triplets, etc.)
}
