% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractInfos.R
\name{transform_digits}
\alias{transform_digits}
\title{Convert digits into genes}
\usage{
transform_digits(combination, genes)
}
\arguments{
\item{combination}{String with binaries (0 and 1) separated by underscores
("_").}

\item{genes}{Vector of genes that were targeted for.}
}
\value{
Vector containing the transformed strings.
}
\description{
Convert the binary digits of into the genes they represent (i.e., match 
target from csv file to genes).
}
