% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Writers.R
\name{write_output_file}
\alias{write_output_file}
\title{Write data to xlsx}
\usage{
write_output_file(
  output_tables,
  conf_mats,
  output_file,
  h2o_table,
  multi_pos,
  shear_table
)
}
\arguments{
\item{output_tables}{List of data frames. Output from create_tables().}

\item{conf_mats}{The computed confusion matrices.}

\item{output_file}{File name (path) of the output file.}

\item{h2o_table}{Data frame with H2O samples created in create_tables().}

\item{multi_pos}{The possible multiple positives.}

\item{shear_table}{Output from compute_shearing_factor.}
}
\value{
No return value. Creates an xlsx file as output_file.
}
\description{
Write finished analysis to a xlsx file
}
