% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TotalHIVDNA.R
\name{compute_total_HIV}
\alias{compute_total_HIV}
\title{Compute total HIV content}
\usage{
compute_total_HIV(df)
}
\arguments{
\item{df}{The data frame for whose column the total HIV content shall be computed}
}
\value{
Data frame df updated with new columns.
}
\description{
Compute the total HIV content. If quadruple positive values are not found in the
data, the following order of multiplets will be used until a substitute is found:
(quadrupel > EnvGagPsi > EnvPsiPol > EnvPsi > EnvGagPol > EnvGag). For more information
check the associated publication.
}
