% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractInfos.R
\name{compute_total_HIV}
\alias{compute_total_HIV}
\title{Compute total HIV content}
\usage{
compute_total_HIV(df, multi_pos)
}
\arguments{
\item{df}{The data frame for whose column the total HIV content shall be computed}

\item{multi_pos}{List of multiple positives. #TODO: check again.}
}
\value{
Data frame df updated with new columns.
}
\description{
Compute the total HIV content as proposed by Rachel Scheck.
}
