% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractInfos.R
\name{get_multipos}
\alias{get_multipos}
\title{Get multiple positive combinations}
\usage{
get_multipos(genes)
}
\arguments{
\item{genes}{List of genes used in the experiment.}
}
\value{
List of possible combinations
}
\description{
Compute the possible combinations of genes based on the genes in an experiment.
}
