% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Writers.R
\name{get_output_sheet}
\alias{get_output_sheet}
\title{Create analysis tables}
\usage{
get_output_sheet(table, multi_pos, grouped_data)
}
\arguments{
\item{table}{Data frame computed in create_tables().}

\item{multi_pos}{The possible multiple positives.}

\item{grouped_data}{The grouped data obtained by define_groups().}
}
\value{
A named vector as blueprint for a xlsx sheet.
}
\description{
Create an output sheet that collects the most important information.
}
\keyword{internal}
